==> scRNA-seq/MULTI-seq Analysis/0.Getting_sample_barcodes.ipynb <==
# Getting the MULTIseq barcodes used for this analysis

Update on 24-11-2022

setwd("/project/ABKlab/Natalia/sc-RNA-seq/single_TLS_project")

library("readxl")

# xlsx files
my_data_48h <- read_excel("./raw_data/221103_morphotype/Multi_BC_Variation_With_Morpho.xlsx", sheet = "48h")
my_data_72h <- read_excel("./raw_data/221103_morphotype/Multi_BC_Variation_With_Morpho.xlsx", sheet = "72h")
my_data_96h <- read_excel("./raw_data/221103_morphotype/Multi_BC_Variation_With_Morpho.xlsx", sheet = "96h")
my_data_120h <- read_excel("./raw_data/221103_morphotype/Multi_BC_Variation_With_Morpho.xlsx", sheet = "120h")

dim(my_data_48h)
head(my_data_48h)

# Adding barcode column (substracted from Multi_BC original column)

# substr(x, start, stop)
my_data_48h$barcodes <- substr(my_data_48h$Multi_BC, 22,29)
my_data_72h$barcodes <- substr(my_data_72h$Multi_BC, 22,29)
my_data_96h$barcodes <- substr(my_data_96h$Multi_BC, 22,29)
my_data_120h$barcodes <- substr(my_data_120h$Multi_BC, 22,29)

head(my_data_48h)

barcodes_48h <- my_data_48h$barcodes
barcodes_72h <- my_data_72h$barcodes
barcodes_96h <- my_data_96h$barcodes
barcodes_120h <- my_data_120h$barcodes

length(barcodes_48h)
length(barcodes_72h)
length(barcodes_96h)
length(barcodes_120h)

# Save an object to a file
# saveRDS(object, file = "my_data.rds")

saveRDS(barcodes_48h, file = "./files/barcodes_48h.rds")
saveRDS(barcodes_72h, file = "./files/barcodes_72h.rds")
saveRDS(barcodes_96h, file = "./files/barcodes_96h.rds")
saveRDS(barcodes_120h, file = "./files/barcodes_120h.rds")
