==> scRNA-seq/MULTI-seq Analysis/1.MULTI-seq_1.ipynb <==
# Reading MULTI-seq raw data 

# MULTIseq follow up for ReadTable filtered

setwd("/project/ABKlab/Natalia/sc-RNA-seq/single_TLS_project")
# devtools::install_github('chris-mcginnis-ucsf/MULTI-seq')
library("deMULTIplex")

cell.id.vec <- read.table("../whitelist/3M-february-2018.txt")
cell.id.vec <- as.vector(cell.id.vec$V1)
str(cell.id.vec)


# Equivalence fastq files with timepoints
# 1: 48h
# 2: 72h
# 3: 96h
# 4: 120h_1
# 5: 120h_2

## Pre-process MULTI-seq sample barcode FASTQs

x <- MULTIseq.preProcess(R1 = './fastq_files/mpimg_L24507-1_JV-DNA-1_S1_R1_001.fastq.gz',
                                  R2 = './fastq_files/mpimg_L24507-1_JV-DNA-1_S1_R2_001.fastq.gz',
                                  cellIDs = cell.id.vec, cell=c(1,16), umi=c(17,28), tag=c(1,8))

readTable_1 <- x

x <- MULTIseq.preProcess(R1 = './fastq_files/mpimg_L24508-1_JV-DNA-2_S2_R1_001.fastq.gz',
                                  R2 = './fastq_files/mpimg_L24508-1_JV-DNA-2_S2_R2_001.fastq.gz',
                                  cellIDs = cell.id.vec, cell=c(1,16), umi=c(17,28), tag=c(1,8))

readTable_2 <- x

x <- MULTIseq.preProcess(R1 = './fastq_files/mpimg_L24509-1_JV-DNA-3_S3_R1_001.fastq.gz',
                                  R2 = './fastq_files/mpimg_L24509-1_JV-DNA-3_S3_R2_001.fastq.gz',
                                  cellIDs = cell.id.vec, cell=c(1,16), umi=c(17,28), tag=c(1,8))

readTable_3 <- x

x <- MULTIseq.preProcess(R1 = './fastq_files/mpimg_L24510-1_JV-DNA-4_S4_R1_001.fastq.gz',
                                  R2 = './fastq_files/mpimg_L24510-1_JV-DNA-4_S4_R2_001.fastq.gz',
                                  cellIDs = cell.id.vec, cell=c(1,16), umi=c(17,28), tag=c(1,8))

readTable_4 <- x

x <- MULTIseq.preProcess(R1 = './fastq_files/mpimg_L24511-1_JV-DNA-5_S5_R1_001.fastq.gz',
                                  R2 = './fastq_files/mpimg_L24511-1_JV-DNA-5_S5_R2_001.fastq.gz',
                                  cellIDs = cell.id.vec, cell=c(1,16), umi=c(17,28), tag=c(1,8))

readTable_5 <- x


# Save the readTables

# Save the tables
saveRDS(readTable_1, file = "./files/readTable_1.rds")
saveRDS(readTable_2, file = "./files/readTable_2.rds")
saveRDS(readTable_3, file = "./files/readTable_3.rds")
saveRDS(readTable_4, file = "./files/readTable_4.rds")
saveRDS(readTable_5, file = "./files/readTable_5.rds")

# Load the readTables

readTable_1 <- readRDS("./files/readTable_1.rds")
readTable_2 <- readRDS("./files/readTable_2.rds")
readTable_3 <- readRDS("./files/readTable_3.rds")
readTable_4 <- readRDS("./files/readTable_4.rds")
readTable_5 <- readRDS("./files/readTable_5.rds")

## Load the barcode files

barcodes_48h <- readRDS("./files/barcodes_48h.rds")
barcodes_72h <- readRDS("./files/barcodes_72h.rds")
barcodes_96h <- readRDS("./files/barcodes_96h.rds")
barcodes_120h <- readRDS("./files/barcodes_120h.rds")

# Filter the readTables by used sample barcodes

# Filtering according to barcodes used in each sample
readTable_1_filtered <- readTable_1[readTable_1$Sample %in% barcodes_48h, ]
readTable_2_filtered <- readTable_2[readTable_2$Sample %in% barcodes_72h, ]
readTable_3_filtered <- readTable_3[readTable_3$Sample %in% barcodes_96h, ]
readTable_4_filtered <- readTable_4[readTable_4$Sample %in% barcodes_120h, ]
readTable_5_filtered <- readTable_5[readTable_5$Sample %in% barcodes_120h, ]

# Save the tables
saveRDS(readTable_1_filtered, file = "./files/readTable_1_filtered.rds")
saveRDS(readTable_2_filtered, file = "./files/readTable_2_filtered.rds")
saveRDS(readTable_3_filtered, file = "./files/readTable_3_filtered.rds")
saveRDS(readTable_4_filtered, file = "./files/readTable_4_filtered.rds")
saveRDS(readTable_5_filtered, file = "./files/readTable_5_filtered.rds")

length(table(readTable_1_filtered$Sample))
length(table(readTable_2_filtered$Sample))
length(table(readTable_3_filtered$Sample))
length(table(readTable_4_filtered$Sample))
length(table(readTable_5_filtered$Sample))

readTable_1 <- readRDS("./files/readTable_1_filtered.rds")
head(readTable_1)


str(readTable_1)
