==> scRNA-seq/MULTI-seq Analysis/2.Filtering_ReadTables.ipynb <==
# Filtering ReadTables based on cell barcode

## ANALYSIS WITH ONLY 50M READS OF THE SAMPLE BARCODE FASTQ FILE ##

setwd("/project/ABKlab/Natalia/sc-RNA-seq/single_TLS_project")

# Reading gene-cell matrix output from Cell Ranger

# install.packages("Matrix")
library(Matrix)
library("deMULTIplex")


## 48h (readTable1)

matrix_dir = "./raw_data/cellranger_results/cr-count_mm10_JV-RNA-1/outs/filtered_feature_bc_matrix/"
barcode.path <- paste0(matrix_dir, "barcodes.tsv.gz")
features.path <- paste0(matrix_dir, "features.tsv.gz")
matrix.path <- paste0(matrix_dir, "matrix.mtx.gz")
mat <- readMM(file = matrix.path)
feature.names = read.delim(features.path,
                           header = FALSE,
                           stringsAsFactors = FALSE)
barcode.names = read.delim(barcode.path,
                           header = FALSE,
                           stringsAsFactors = FALSE)
colnames(mat) = barcode.names$V1
rownames(mat) = feature.names$V1



cellbarcode1 <- colnames(mat)
head(cellbarcode1)

cellbarcode1 <- colnames(mat)
cellbarcode1 <- substr(cellbarcode1,1,nchar(cellbarcode1)-2)
head(cellbarcode1)



# Read table from MULTIplex first step
library("deMULTIplex")
readTable_1_filtered <- readRDS(file = "files/readTable_1_filtered.rds", refhook = NULL)


head(readTable_1_filtered)
class(readTable_1_filtered)
str(readTable_1_filtered)
# filter readTable with Cell Ranger cell barcodes that are kept.
readTable_1_filtered <- readTable_1_filtered[readTable_1_filtered$Cell %in% cellbarcode1 ,]


# save readTable_filtered
saveRDS(readTable_1_filtered, file = "files/readTable_1_filtered.rds")
saveRDS(cellbarcode, file = "files/cellbarcode_1.rds")


## 72h (readTable2)

matrix_dir = "./raw_data/cellranger_results/cr-count_mm10_JV-RNA-2/outs/filtered_feature_bc_matrix/"
barcode.path <- paste0(matrix_dir, "barcodes.tsv.gz")
features.path <- paste0(matrix_dir, "features.tsv.gz")
matrix.path <- paste0(matrix_dir, "matrix.mtx.gz")
mat <- readMM(file = matrix.path)
feature.names = read.delim(features.path,
                           header = FALSE,
                           stringsAsFactors = FALSE)
barcode.names = read.delim(barcode.path,
                           header = FALSE,
                           stringsAsFactors = FALSE)
colnames(mat) = barcode.names$V1
rownames(mat) = feature.names$V1


cellbarcode2 <- colnames(mat)
cellbarcode2 <- substr(cellbarcode2,1,nchar(cellbarcode2)-2)
head(cellbarcode2)

# Read table from MULTIplex first step
library("deMULTIplex")
readTable_2_filtered <- readRDS(file = "files/readTable_2_filtered.rds", refhook = NULL)


head(readTable_2_filtered)
class(readTable_2_filtered)
str(readTable_2_filtered)
# filter readTable with Cell Ranger cell barcodes that are kept.
readTable_2_filtered <- readTable_2_filtered[readTable_2_filtered$Cell %in% cellbarcode2 ,]

# save readTable_filtered
saveRDS(readTable_2_filtered, file = "files/readTable_2_filtered.rds")
saveRDS(cellbarcode2, file = "files/cellbarcode_2.rds")



## 96h (readTable3)

matrix_dir = "./raw_data/cellranger_results/cr-count_mm10_JV-RNA-3/outs/filtered_feature_bc_matrix/"
barcode.path <- paste0(matrix_dir, "barcodes.tsv.gz")
features.path <- paste0(matrix_dir, "features.tsv.gz")
matrix.path <- paste0(matrix_dir, "matrix.mtx.gz")
mat <- readMM(file = matrix.path)
feature.names = read.delim(features.path,
                           header = FALSE,
                           stringsAsFactors = FALSE)
barcode.names = read.delim(barcode.path,
                           header = FALSE,
                           stringsAsFactors = FALSE)
colnames(mat) = barcode.names$V1
rownames(mat) = feature.names$V1


cellbarcode3 <- colnames(mat)
cellbarcode3 <- substr(cellbarcode3,1,nchar(cellbarcode3)-2)
head(cellbarcode3)

# Read table from MULTIplex first step
library("deMULTIplex")
readTable_3_filtered <- readRDS(file = "files/readTable_3_filtered.rds", refhook = NULL)


head(readTable_3_filtered)
class(readTable_3_filtered)
str(readTable_3_filtered)
# filter readTable with Cell Ranger cell barcodes that are kept.
readTable_3_filtered <- readTable_3_filtered[readTable_3_filtered$Cell %in% cellbarcode3 ,]

# save readTable_filtered
saveRDS(readTable_3_filtered, file = "files/readTable_3_filtered.rds")
saveRDS(cellbarcode3, file = "files/cellbarcode_3.rds")



## 120h_1 (readTable4)

matrix_dir = "./raw_data/cellranger_results/cr-count_mm10_JV-RNA-4/outs/filtered_feature_bc_matrix/"
barcode.path <- paste0(matrix_dir, "barcodes.tsv.gz")
features.path <- paste0(matrix_dir, "features.tsv.gz")
matrix.path <- paste0(matrix_dir, "matrix.mtx.gz")
mat <- readMM(file = matrix.path)
feature.names = read.delim(features.path,
                           header = FALSE,
                           stringsAsFactors = FALSE)
barcode.names = read.delim(barcode.path,
                           header = FALSE,
                           stringsAsFactors = FALSE)
colnames(mat) = barcode.names$V1
rownames(mat) = feature.names$V1


cellbarcode4 <- colnames(mat)
cellbarcode4 <- substr(cellbarcode4,1,nchar(cellbarcode4)-2)
head(cellbarcode4)

# Read table from MULTIplex first step
library("deMULTIplex")
readTable_4_filtered <- readRDS(file = "files/readTable_4_filtered.rds", refhook = NULL)


head(readTable_4_filtered)
class(readTable_4_filtered)
str(readTable_4_filtered)
# filter readTable with Cell Ranger cell barcodes that are kept.
readTable_4_filtered <- readTable_4_filtered[readTable_4_filtered$Cell %in% cellbarcode4 ,]

# save readTable_filtered
saveRDS(readTable_4_filtered, file = "files/readTable_4_filtered.rds")
saveRDS(cellbarcode4, file = "files/cellbarcode_4.rds")

## 120h_2 (readTable5)

matrix_dir = "./raw_data/cellranger_results/cr-count_mm10_JV-RNA-5/outs/filtered_feature_bc_matrix/"
barcode.path <- paste0(matrix_dir, "barcodes.tsv.gz")
features.path <- paste0(matrix_dir, "features.tsv.gz")
matrix.path <- paste0(matrix_dir, "matrix.mtx.gz")
mat <- readMM(file = matrix.path)
feature.names = read.delim(features.path,
                           header = FALSE,
                           stringsAsFactors = FALSE)
barcode.names = read.delim(barcode.path,
                           header = FALSE,
                           stringsAsFactors = FALSE)
colnames(mat) = barcode.names$V1
rownames(mat) = feature.names$V1


cellbarcode5 <- colnames(mat)
cellbarcode5 <- substr(cellbarcode5,1,nchar(cellbarcode5)-2)
head(cellbarcode5)

# Read table from MULTIplex first step
library("deMULTIplex")
readTable_5_filtered <- readRDS(file = "files/readTable_5_filtered.rds", refhook = NULL)


head(readTable_5_filtered)
class(readTable_5_filtered)
str(readTable_5_filtered)
# filter readTable with Cell Ranger cell barcodes that are kept.
readTable_5_filtered <- readTable_5_filtered[readTable_5_filtered$Cell %in% cellbarcode5 ,]

# save readTable_filtered
saveRDS(readTable_5_filtered, file = "files/readTable_5_filtered.rds")
saveRDS(cellbarcode5, file = "files/cellbarcode_5.rds")

# MULTIseq alignments

This is the way it is done but better to paralelize the job

setwd("/project/ABKlab/Natalia/sc-RNA-seq/single_TLS_project")

library("deMULTIplex")

cell.id.vec <- read.table("../whitelist/3M-february-2018.txt")
cell.id.vec <- as.vector(cell.id.vec$V1)
str(cell.id.vec)

barcodes_48h <- readRDS("./files/barcodes_48h.rds")
barcodes_72h <- readRDS("./files/barcodes_72h.rds")
barcodes_96h <- readRDS("./files/barcodes_96h.rds")
barcodes_120h <- readRDS("./files/barcodes_120h.rds")

readTable_1_filtered <- readRDS("./files/readTable_1_filtered.rds")
readTable_2_filtered <- readRDS("./files/readTable_2_filtered.rds")
readTable_3_filtered <- readRDS("./files/readTable_3_filtered.rds")
readTable_4_filtered <- readRDS("./files/readTable_4_filtered.rds")
readTable_5_filtered <- readRDS("./files/readTable_5_filtered.rds")

head(cellbarcode5)
length(cellbarcode5)

head(readTable_1_filtered)
dim(readTable_1_filtered)

head(cellbarcode1)
length(cellbarcode1)

head(barcodes_48h)
length(barcodes_48h)

library("deMULTIplex")

## Perform MULTI-seq sample barcode alignment
bar.table_1 <- MULTIseq.align(readTable_1_filtered, cellbarcode1, barcodes_48h)
saveRDS(bar.table_1, file = "./files/bartable_1.rds")

bar.table_2 <- MULTIseq.align(readTable_2_filtered, cellbarcode2, barcodes_72h)
saveRDS(bar.table_2, file = "./files/bartable_2.rds")

bar.table_3 <- MULTIseq.align(readTable_3_filtered, cellbarcode3, barcodes_96h)
saveRDS(bar.table_3, file = "./files/bartable_3.rds")

bar.table_4 <- MULTIseq.align(readTable_4_filtered, cellbarcode4, barcodes_120h)
saveRDS(bar.table_4, file = "./files/bartable_4.rds")

bar.table_5 <- MULTIseq.align(readTable_5_filtered, cellbarcode5, barcodes_120h)
saveRDS(bar.table_5, file = "./files/bartable_5.rds")


# Demultiplexing 48h

## Visualize barcode space
bar.tsne_1 <- barTSNE(bar.table_1[,1:3]) 
## Note: Exclude columns 97:98 (assuming 96 barcodes were used) which provide total barcode UMI counts for each cell. 

pdf("./outputs/MULTIseq/bar.tsne_1.check.pdf")
for (i in 3:ncol(bar.tsne_1)) {
  g <- ggplot(bar.tsne_1, aes(x = TSNE1, y = TSNE2, color = bar.tsne_1[,i])) +
    geom_point() +
    scale_color_gradient(low = "black", high = "red") +
    ggtitle(colnames(bar.tsne_1)[i]) +
    theme(legend.position = "none") 
  print(g)
}
dev.off()

print(g)

head(bar.table_5)
